Refactor application of APO defaults
The controller was doing too much. The service still doing a lot... but the data model is what it is currently. We could consider further rounds of refactoring in the future.
